% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpn.R
\name{rpn}
\alias{rpn}
\title{Reverse Polish Notation converter/interpreter.}
\usage{
rpn(rpn.expr, reverse = TRUE, eval = TRUE, clean = TRUE, vars = list(),
  ops = list())
}
\arguments{
\item{rpn.expr}{[\code{character}]\cr
Character vector representing the the expression in (reverse) polish notation.}

\item{reverse}{[\code{logical(1)}]\cr
Is \code{rpn.expr} in reverse polish notation or polish notation?
Default is \code{TRUE}, i.e., reverse polish notation.}

\item{eval}{[\code{logical(1)}]\cr
Shall the expression be evaluated?
Default is \code{TRUE}.}

\item{clean}{[\code{logical(1)}]\cr
Should braces be removed from \code{rpn.expr} before interpretation?
Default is \code{TRUE}.}

\item{vars}{[\code{list}]\cr
Named list of variables which are used in the \code{rpn.expr}.
Default is the empty list.}

\item{ops}{[\code{list}]\cr
Optional named list of operators. If non-empty, the list must be of key-value
type, i.e., the key is the name of the function and the value is the length
of the parameter vector the corresponding function expects.
The default is the empty list. In this case the only operators which can be
evaluated are \dQuote{+}, \dQuote{-}, \dQuote{*} and \dQuote{/}.}
}
\value{
Evaluated expression or expression in infix notation.
}
\description{
Expects an expression in (reverse) polish notation as a character vector and
returns the expression in infix notation and the evaluated expression.
}
\details{
This is a pure R implementation.
}
\examples{
r = c("4", "6", "*", "6", "+")
rpn(r)
rpn(r, eval = FALSE)

r = c("x", "6", "*", "6", "+")
rpn(r, eval = TRUE, vars = list(x = 4))
rpn(r, eval = FALSE)

}
\references{
...
}

